Modify: 각 디렉토리 진입점 파일명 index.jsx 수정완료, 파일구조 일부수정
