+++ client/src/Pages/Detail/components/DetailInfo/shared/InfoContainer.jsx
