+++ client/src/Pages/Detail/components/DetailModal/index.jsx
