+++ client/src/Pages/Register/components/RegisterAbove/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import AboveTitle from './AboveTitle'
+import AboveSocial from './AboveSocial'
+
+const RegisterAbove = () => {
+  return (
+    <>
+      <AboveTitle />
+      <AboveSocial />
+    </>
+  )
+}
+
+export default RegisterAbove
