+++ client/src/Pages/Register/index.jsx
@@ -1,17 +1,17 @@
 import React from 'react'
 import styled from 'styled-components'
-import RegisterTitle from './components/RegisterTitle'
-import RegisterSocial from './components/RegisterSocial'
-import RegisterDivider from './components/RegisterDivider'
-import RegisterForm from './components/RegisterForm/RegisterForm';
+import RegisterAbove from './components/RegisterAbove';
+import RegisterForm from './components/RegisterForm';
 
 const Register = () => {
   return (
     <RegisterPage>
       <RegisterWrapper>
-        <RegisterTitle />
-        <RegisterSocial />
-        <RegisterDivider />
+        <RegisterAbove />
+        <RegisterDivider>
+          <div className="divider-line" />
+          <span>OR</span>
+        </RegisterDivider>
         <RegisterForm />
       </RegisterWrapper>
     </RegisterPage>
@@ -29,4 +29,26 @@ const RegisterWrapper = styled.main`
   margin: 0 auto;
 `;
 
-export default Register
+const RegisterDivider = styled.div`
+  position: relative;
+  margin: 40px 0;
+  
+  .divider-line {
+    width: 100%;
+    height: 1px;
+    background: ${({ theme }) => theme.gray0C};
+  }
+
+  span {
+    position: absolute;
+    top: -7px;
+    left: 50%;
+    transform: translateX(-50%);
+    padding: 0 20px;
+    background: #ffffff;
+    color: ${({ theme }) => theme.gray2};
+    font-size: 14px;
+  }
+`;
+
+export default Register;
